requirements.txt
fastapi>=0.115
uvicorn[standard]
pydantic>=2.0
pydantic-settings>=2.0
sqlite-vec
# embedding走外部API（智谱embedding-3），不装本地模型
httpx
python-dotenv
pytest
pytest-asyncio
